email_processor/pyproject.toml
[project]
name = "email-processor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.1",
    "click>=8.2.1",
    "google-adk>=1.11.0",
    "google-auth-oauthlib>=1.2.2",
    "requests>=2.32.5",
    "rich>=14.1.0",
    "uvicorn>=0.35.0",
]
